use palette::Gradient for wheel colors
this is a pretty basic impl that just maps our existing RYB points to a gradient.
improves on the status quo a little bit by getting the palette crate into the mix
and allowing aribrary number of divisions in the color wheel, although that still
needs to be wired up to ColorWheel props, etc. before it will be useful