wip: use web renderer
need to wire up the midi crate and test it in wasm...
it should "just work" and use the WebMidi backend, I think.
If so, maybe worth making web the main render target.
Seems like feedback loops might be faster.

cant seem to depend on desktop and build for wasm at
the same time. so if we want both, we need to make
little wrapper crates for each platform.